fix: Provide individual exports for runtime APIs
