+++ packages/wxt/package.json
@@ -3,68 +3,7 @@
   "type": "module",
   "version": "0.19.16",
   "description": "Next gen framework for developing web extensions",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/wxt-dev/wxt.git"
-  },
-  "homepage": "https://wxt.dev",
-  "keywords": [
-    "vite",
-    "chrome",
-    "web",
-    "extension",
-    "browser",
-    "bundler",
-    "framework"
-  ],
-  "author": {
-    "name": "Aaron Klinker",
-    "email": "aaronklinker1+wxt@gmail.com"
-  },
   "license": "MIT",
-  "files": [
-    "bin",
-    "dist"
-  ],
-  "bin": {
-    "wxt": "./bin/wxt.mjs",
-    "wxt-publish-extension": "./bin/wxt-publish-extension.cjs"
-  },
-  "module": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.mjs"
-    },
-    "./client": {
-      "types": "./dist/client/index.d.ts",
-      "default": "./dist/client/index.mjs"
-    },
-    "./sandbox": {
-      "types": "./dist/sandbox/index.d.ts",
-      "default": "./dist/sandbox/index.mjs"
-    },
-    "./browser": {
-      "types": "./dist/browser.d.ts",
-      "default": "./dist/browser.mjs"
-    },
-    "./testing": {
-      "types": "./dist/testing/index.d.ts",
-      "default": "./dist/testing/index.mjs"
-    },
-    "./storage": {
-      "types": "./dist/storage.d.ts",
-      "default": "./dist/storage.mjs"
-    },
-    "./vite-builder-env": {
-      "types": "./dist/vite-builder-env.d.ts"
-    },
-    "./modules": {
-      "types": "./dist/modules.d.ts",
-      "default": "./dist/modules.mjs"
-    }
-  },
   "scripts": {
     "wxt": "tsx src/cli/index.ts",
     "build": "buildc -- unbuild",
@@ -142,6 +81,119 @@
   "peerDependenciesMeta": {
     "@types/chrome": {
       "optional": true
+    }
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/wxt-dev/wxt.git"
+  },
+  "homepage": "https://wxt.dev",
+  "keywords": [
+    "vite",
+    "chrome",
+    "web",
+    "extension",
+    "browser",
+    "bundler",
+    "framework"
+  ],
+  "author": {
+    "name": "Aaron Klinker",
+    "email": "aaronklinker1+wxt@gmail.com"
+  },
+  "files": [
+    "bin",
+    "dist"
+  ],
+  "bin": {
+    "wxt": "./bin/wxt.mjs",
+    "wxt-publish-extension": "./bin/wxt-publish-extension.cjs"
+  },
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.mjs"
+    },
+    "./client": {
+      "types": "./dist/client/index.d.ts",
+      "default": "./dist/client/index.mjs"
+    },
+    "./client/useAppConfig": {
+      "types": "./dist/client/app-config.d.ts",
+      "default": "./dist/client/app-config.mjs"
+    },
+    "./client/injectScript": {
+      "types": "./dist/client/inject-script.d.ts",
+      "default": "./dist/client/inject-script.mjs"
+    },
+    "./client/ContentScriptContext": {
+      "types": "./dist/client/content-script-context.d.ts",
+      "default": "./dist/client/content-script-context.mjs"
+    },
+    "./client/ui/types": {
+      "types": "./dist/client/ui/types.d.ts",
+      "default": "./dist/client/ui/types.mjs"
+    },
+    "./client/ui/createIntegratedUi": {
+      "types": "./dist/client/ui/create-integrated-ui.d.ts",
+      "default": "./dist/client/ui/create-integrated-ui.mjs"
+    },
+    "./client/ui/createShadowRootUi": {
+      "types": "./dist/client/ui/create-shadow-root-ui.d.ts",
+      "default": "./dist/client/ui/create-shadow-root-ui.mjs"
+    },
+    "./client/ui/createIframeUi": {
+      "types": "./dist/client/ui/create-iframe-ui.d.ts",
+      "default": "./dist/client/ui/create-iframe-ui.mjs"
+    },
+    "./sandbox": {
+      "types": "./dist/sandbox/index.d.ts",
+      "default": "./dist/sandbox/index.mjs"
+    },
+    "./sandbox/defineAppConfig": {
+      "types": "./dist/sandbox/define-app-config.d.ts",
+      "default": "./dist/sandbox/define-app-config.mjs"
+    },
+    "./sandbox/defineBackground": {
+      "types": "./dist/sandbox/define-background.d.ts",
+      "default": "./dist/sandbox/define-background.mjs"
+    },
+    "./sandbox/defineContentScript": {
+      "types": "./dist/sandbox/define-content-script.d.ts",
+      "default": "./dist/sandbox/define-content-script.mjs"
+    },
+    "./sandbox/defineUnlistedScript": {
+      "types": "./dist/sandbox/define-unlisted-script.d.ts",
+      "default": "./dist/sandbox/define-unlisted-script.mjs"
+    },
+    "./sandbox/defineWxtPlugin": {
+      "types": "./dist/sandbox/define-wxt-plugin.d.ts",
+      "default": "./dist/sandbox/define-wxt-plugin.mjs"
+    },
+    "./browser": {
+      "types": "./dist/browser/index.d.ts",
+      "default": "./dist/browser/index.mjs"
+    },
+    "./browser/chrome": {
+      "types": "./dist/browser/chrome.d.ts",
+      "import": "./dist/browser/chrome.mjs"
+    },
+    "./testing": {
+      "types": "./dist/testing/index.d.ts",
+      "default": "./dist/testing/index.mjs"
+    },
+    "./storage": {
+      "types": "./dist/storage.d.ts",
+      "default": "./dist/storage.mjs"
+    },
+    "./vite-builder-env": {
+      "types": "./dist/vite-builder-env.d.ts"
+    },
+    "./modules": {
+      "types": "./dist/modules.d.ts",
+      "default": "./dist/modules.mjs"
     }
   }
 }
